Uploaded the latest version of lab-1 by replacing the original code with assembly one.
Directly compare the input with 0 and 1 in the beginning.
Use other registers (R1, R2, R3) to calculate the sequence loop.
Operating the loop by using a label and counting the input with subtraction.
Edit the two labels in the bottom of the original code to fit the latest code a$